Clean up and organize repository structure
Reorganize documentation:
- Move all docs to docs/ folder with subfolders
- Create docs/wireframes/ for HTML mockups
- Create docs/reports/ for progress tracking
- Move architecture docs to logical locations

Update README:
- Complete rewrite for hackathon demo
- Add quick start guide
- Document all Kaiz dashboard pages
- Include troubleshooting section
- Add clear project structure

Security improvements:
- Ensure .gitignore covers all sensitive files
- Add security warnings in README

File organization:
- All HTML wireframes in docs/wireframes/
- All progress reports in docs/reports/
- Main docs in docs/ root
- Clean root directory

+++ scripts/execute-demo-payroll.ts
@@ -110,7 +110,7 @@ async function main() {
     args: [treasuryAccount.address],
   });
 
-  const balanceUSDC = formatEther(balance);
+  const balanceUSDC = formatEther(balance as bigint);
   console.log(`   Balance: ${balanceUSDC} USDC\n`);
 
   if (parseFloat(balanceUSDC) < 100) {
